submit: 4.21*: Blog List Expansion, step 9
- Change the delete blog operation so that a blog can be deleted only by the user who added it...

+++ controllers/blogs.js
@@ -43,15 +43,29 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  try {
-    const deleted = await Blog.findByIdAndDelete(request.params.id);
-    if (deleted) {
-      response.status(204).end();
-    } else {
-      response.status(404).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  console.log('blog.user', blog.user.toString());
+  console.log('user.id', user.id);
+
+  if (blog.user.toString() === user.id) {
+    try {
+      const deleted = await Blog.findByIdAndDelete(request.params.id);
+      if (deleted) {
+        response.status(204).end();
+      } else {
+        response.status(404).end();
+      }
+    } catch (err) {
+      if (err.name === 'CastError') response.status(400).end().json({ error: 'Invalid blog ID' });
     }
-  } catch (err) {
-    if (err.name === 'CastError') response.status(400).end().json({ error: 'Invalid blog ID' });
+  } else {
+    return response.status(401).json({ error: 'unauthroized to delete' });
   }
 });
